tty/8250_early: Don't truncate last character of the options
the earlier change to use strlcpy uncovered a bug in the options
argument length calculation causing last character of the options
string to be truncated by strlcpy. This makes the actual console
to be configured with incorrect baudrate when specifying the console
using console=uart,... syntax.

Bug symptom seen in kernel log output:

Kernel command line: console=uart,mmio,0x90000000,115200
Early serial console at MMIO 0x90000000 (options '11520')

which then results in a invalid baud rate 11520 instead of the
expected 115200 when the console is switched to ttyS0 later
in the boot process.

Signed-off-by: Henrik Nordstrom <henrik@henriknordstrom.net>